+++ tests/polars/functions/test_convert_excel_date_polars.py
@@ -1,0 +1,11 @@
+import polars as pl
+
+import janitor.polars  # noqa: F401
+
+
+def test_convert_excel_date():
+    df = pl.DataFrame({"dates": [42580.3333333333]})
+
+    expression = pl.col("dates").janitor.convert_excel_date().alias("dd")
+    expression = df.with_columns(expression).get_column("dd")
+    assert expression.dtype.is_temporal() is True